Don't call invalidate when resetting the QSortFilterProxyModel.
The invalidate connection is made before connecting the source model,
which means it gets called first.

If a second proxy model is used with it, the sourceLayoutAboutToBeChanged
of the second proxy is therefore called before its sourceReset slot, but
after the persistent indexes of the first model have been invalidated.

The invalidate call is not needed because clear_mappings is called
when the source model is reset anyway.

Merge-request: 416
Reviewed-by: Olivier Goffart <ogoffart@trolltech.com>