Make Qt use Cocoa by default on Mac OS X
The "-cococa" option is now deprecated, "-carbon" is added.

This is mostly a behind-the-scenes change, but has some implications:

If you are developing on Snow Leopard Qt vill be built in x86_64
mode by default. This can be overridden by configuring with "-arch x86"

Qt for Cocoa will not build on 10.4 Tiger, nor will Cocoa-based
binaries deploy back to that platform. Configure with "-carbon"
if you want to target 10.4.