Don't write out fo:word-spacing if its the default value.
the property fo:word-spacing is still not in an official ODF spec so
a strict schema check will fail in any doc that uses it.  Lets make sure
we limit the writing of that property to only those usecases where the
property was really set by the user.

Task-number: QTBUG-5725