Cocoa: problems quitting application when not using qApp::exec
This problem has been around before. At that time it had to do
with what part of the application received the kEAQuit apple event.
In Qt-4.7 we ensure that NSApplication gets run at least once
before we do any other event processing in Qt. That way the menu bar
gets fully initialized. The backside of the coin is that NSApplication
will not let the quit menu send a terminate action to the menu nibs file
owner (qmenuloader_mac.mm) when a modal window is showing on screen.
This is normally expected behaviour, but in Qt we allow the user to quit
when just a single window/dialog is showing on screen.
This fix remaps the quit items target from files owner to first responder.
That way we can catch the terminate at the window/responder beeing that,
and terminate as needed.

Task-number: QTBUG-7197
Reviewed-by: Prasanth