Fix bugs evidenced by change to keep flush errors on QFile::close
When closing a file, the contents of the write buffer should be emptied
even in the case where flush fails, so it doesn't leak to subsequent
files opened through the same instance. This is inline with what would
happen with native close on a buffered device.

Also changed the resource file engine to succeed on flush. Since all
writes fail there, logically it's write buffer is empty and flush should
succeed. This keeps auto-tests happy :-)

tst_QFile::fullDisk auto-test extended to ensure re-opening QFile does
not keep the write buffer alive.

Reviewed-by: Thiago Macieira