Fixes:    Setting any style sheet breaks checkbox positioning
Task:     QTBUG-7198
RevBy:    ogoffart
Details:
It was impossible to override the checkbox positioning in a custom
style if any style sheet was set on the application of widget.
This was because the style sheet never passed control to the
base style. We now fix it by checking if the style sheet has
the appropriate style rules.