Obsolete QDirModel and remove it from the doc.
QFileSystemModel provides the same features as QDirModel but in a
efficient way so no need to keep QDirModel around.
This commit mark as obsolete QDirModel. Examples has been ported to
use QFileSystemModel instead. QCompleter also gain the support of
QFileSystemModel to complete nicely a line edit.

This commit also add a new signal in QFileSystemModel when the directory
is loaded, i.e. the thread has finished to read the current dir.

Task-number:QTBUG-3884
Reviewed-by:ogoffart
Reviewed-by:gabi
Reviewed-by:mbm