Implemented audio effects capability querying in Phonon MMF backend
This patch addresses the following deficiencies in the existing
implementation of audio effects:

1. Native effect objects (e.g. CAudioEqualizer, CBassBoost etc)
were created frequently, just in order to check whether a given
effect is supported, or retrieve the list of parameters which it
requires.  Although this is in part due to a deficiency in the
S60 audio effects API (it doesn't have a 'capability query' concept),
it can be improved by using caching.  This patch introduces a
singleton EffectFactory object, which lazily initializes a data
structure containing information about support and parameters.

2. In order to either query effect support, the native effect object
ultimately needs to access a DevSound instance.  Previously, the
effect object was provided with a CMdaAudioPlayerUtility object.
If this player utility has not loaded an MMF controller plugin
*before* the effects object makes any calls to the player utility,
incorrect results will be returned.  This was observed in the
previous code.  For querying, we don't actually need to load an
MMF controller; instead, we would like to directly provide a
CMMFDevSound instance to the effect object.  Unfortunately, this
API is not available in public S60 SDKs, so we must use the next
lowest interface available, namely CMdaAudioOutputStream.
By making this change, this patch ensures that support and parameter
queries made via the EffectFactory will return the correct result.

At this point, however, effect settings made via the Phonon API
are not actually applied to the audio output.  Fixing this will
require notifying the audio effects backend nodes of state changes
in the MediaObject.

Task-number: QTBUG-4659
Reviewed-by: Espen Riskedal