[syncqt] Make -separate-module work with WebKit's directory structure
Changed the syntax to allow for the situation where the .pro file is in
a different directory than the header files to install. In this case the
resulting Makefile is in WebCore/Makefile, so the references in headers.pri
to for example qwebview.h has to be ../WebKit/qt/Api/qwebview.h.

This doesn't affect any existing syncqt operations inside of Qt.

Reviewed-by: Trust me