Modified effect parameter handling to improve user experience
This change is to work around a limitation in the Phonon::EffectWidget
class.  This widget only displays sliders for parameters with numeric
values if the variant type of the parameter is QReal and the range is
exactly -1.0 to +1.0; otherwise, a spinbox is displayed. This is
rather inconvenient for many effects, such as the audio equalizer, for
which a slider is a much more natural UI control.

The MMF backend therefore reports the type of numeric parameters to
be QReal, and the range to be -1.0 to +1.0.  Internally, the integer
range for the parameter is stored.  Changes to the parameter value are
converted from the client-side, floating point representation to the
internal, integer representation.

Task-number: QTBUG-4659
Reviewed-by: Espen Riskedal