Add an autorelease pool to QApplication instance.
This will catch releases done between creating the
QApplication object and calling QApplication::exec.
(The Cocoa event loop creates release pools for us
when it's running.)

The memory used by the objects stored in this pool
won't be released until the QApplication instance
is deleted at application exit. Because of this it
still makes sense to still add local release pools
when doing potentially large allocations.