Cocoa: the menu bar is sometimes faulty disabled
In Qt, we allow the menu bar to be enabled if we only have one
exec-ing dialog on screen. That way you can reach the quit menu item to
close down the application. But in order for this to work, some overhaul
of qmenu_mac.mm was need to teach it these rules.