doc: This bug wasn't really a bug; it just required clarification.
The reported bug was actually not a bug. It is ok to pass the enum
value as a string here, because the enumeration type has been
registered with the meta-object system with the Q_ENUMS()
macro. However, I have added a bit of text to clarify things a bit.

Task-number:  	 QTBUG-7158
(cherry picked from commit cd0772335cc9bcedd582cacc834bf4609ca650b0)