File dialog mode is not correctly updated in Mac (Cocoa).
QFileDialog shows a customized version of the native file dialog
on Mac. When user changes the fileMode, the corresponding
options needs to be updated in the native dialog as well. Some
options like mode, filters, button text, dialog title etc needs
to be updated for this.

This patch also address the changing of 'acceptMode'. When user
changes this, for e.g. the native dialog needs to switch from an
open panel to save panel. It is easier to recreate the internal
QNSOpenSavePanelDelegate than changing options individually.

Task-number: QTBUG-7086
Reviewed-by: Richard Moe Gustavsen