Use MMP keyword, not compiler flags, to update FPU option on Symbian
This avoids two sets of --fpu flags being passed on the compiler command
line, so we don't have to worry about which takes precedence (especially
since it varies between RVCT and GCC).

Putting ARMFPU in the MMP file means that the toolchain will replace its
default option with the one we specify instead.

Reviewed-by: Aleksandar Sasha Babic