Replace the truncate function with fuzzierCompare().
truncate did a qRound of the two numbers, which could cause two numbers
that were close to become far away (e.g. (0.000049 and 0.000051 would
become 0 and 1 respectively)
fuzzierCompare allows less precision than qFuzzyCompare. This is because
the anchor layout engine can sometimes do quite a lot of calculations,
and the error will be larger than what the qFuzzyCompare checks for.
Thus, the factor in fuzzierCompare is not chosen due to mathematical
proof but rather it is just chosen by what is acceptable.
(actually it can still be larger and still be acceptable)