QListView in icon view mode, drop enabled items wouldn't receive anything
QIconModeViewBase::filterDropEvent() was only moving the items around in the
view without ever checking whether the cursor was over a drop enabled
item. Now it does and returns false if it's the case. As a consequence,
QAbstractItemView::dropEvent() gets called. No auto-test since it's a drag &
drop related task.

Reviewed-by: janarve
Task-number: QTBUG-6848