Cococa: QDialogs stays on top when application is not active
At one point in history, we decided that QDialogs should stack just
above normal windows. This patch reverts this behaviour, as cocoa
gets confused if we stack a dialog at this level without telling it
to be modal. So instead, we let cocoa set the modal window level when
we create a modal session, and also let cocoa restore the level to
what it were when the sessions ends. That way, we avoid doing the work
that cocoa is allready designed to do.

Task-number: QTBUG-5057
Reviewed-by: msorvig