Implement custom QML slot invokation logic
Previously QML was inefficiently forwarding the task of invoking
Qt slots to QScript.  QML does not implement the more advanced
argument coercian of QScript and does not support method overloading.
These two features are only needed to support legacy C++ classes
(of which QML has none), and are not worth the perf cost to support.