Symbian RnD SDK version pf_5250, wk48 and later (e.g., for ivalo target) has changed /epoc32/include layout, including a subdirectory called "platform".
Since that particular directory is used as test for detecting SDKs with "New SF structure", the test returns with false result in case of pf_5250 (which has old SF structure).

Change was made to test for a more obscure directory "mw" that appears the be present in new structure only.

Merge-request: 2281
Signed-off-by: axis <qt-info@nokia.com>