Fix fromSymbian*() image conversion functions.
Do not explicitly create instances of QS60PixmapData in this function
because it is possible that we are using another graphics system. The
fromNativeType() function is virtual so the graphics system will create
the right version of the pixmap data for us and ensure that we get into
the right implementation.

Reviewed-by: Aleksandar Sasha Babic