Designer crashes when previewing QMainWindow with native Toolbar on Mac
The crash happens when a QMainWindow with a native toolbar is reparented.
While re-parenting, the QToolBar triggers a new setParent() with the
MacWindowToolBarButtonHint. This messes up the internal state, hence the
crash. To avoid re-entering setParent() for the QMainWindow, this hint
is set depending on the unifiedTitleAndToolBarOnMac property.

This patch also fixes a related issue in Cocoa, we need to remove the
reused NSToolbar from the old window while recreating the QMainWindow.

Task-number: QTBUG-7162
Reviewed-by: Richard Moe Gustavsen