Return some fake data if we can't handle the network request.
Since we are the only handler for network request and cannot forward the
request to our parent (e.g. unwanted net access), we work arround the
case we can't handle it. We do create a reply with dummy data which will
prevent some blocking and allows further processing of requested data.

Task-number: QTBUG-7427
Reviewed-by: kh