Optimized QLocale to access system locale on demand.
Changed the initialization of the system locale to make construction of the
QLocale object as lightweight as possible. So now the default contructor just
creates a QLocale and QSystemLocale objects, but doesn't try to fill the cache
in the latter with data from the system and postpones it until it is actually
requested (most applications create QLocale objects on the stack and might not
even use the data from the system locale, so we don't need to initialize system
locale right away).

This is an improved version of an already reverted change
8911ed8bfe7f918b93c758f9b5d93274b37739e6,
8911ed8bfe7f918b93c758f9b5d93274b37739e6 and
8911ed8bfe7f918b93c758f9b5d93274b37739e6.

Reviewed-by: Thiago Macieira