SSL Certificate: don't crash when the END CERTIFICATE line ends without CRLF
If the file/data ends in the END CERTIFICATE line without a newline,
the certificate is still valid. If it's followed by anything other
than a newline, then it's no longer valid.

Also add another test for the BEGIN CERTIFICATE ending without
newline, to ensure we don't crash there either.

Reviewed-By: Peter Hartmann