Menubar and dock disappear after hiding a fullscreen widget on Cocoa.
The problem here was exiting the full screen mode. In Cocoa we don't
activate windows as we do in Carbon, therefore we were not exiting from
the full screen mode. This patch adds a check when hiding a window, if
the window is in full screen mode then we go through the list of top
level windows checking if there are any other visible and not-minimized
windows that are also in full screen mode. If none if found, the we exit
the full screen mode.

Task-number: QTBUG-7312
Reviewed-by: Prasanth