Made characters in a password field briefly visible while typing.
This was done by intercepting key events with text in them, and
temporarily submit them as preedit text instead of real input text.

Currently it does not work in WebKit, but that is because WebKit
hides preedit text as well, which is a bug of its own.

RevBy:    Simon Hausmann
Autotest: Manual testing went fine