Fix y-inverted pixmaps properly.
There is a lot of code depending on that pixmaps are flipped
upside down in the gl graphicssystem, so toggling this requires
extensive testing. Since we're anyway questioning the relevance
of this feature (compared to raster + GL viewport) its simply not
worth the effort to fix it properly right now.

Revert "Fixed y-inverted pixmaps on N900."

This reverts commit 57473d5d2a7bd6ae3117f61ff29264a1b790bb01.