The declaration of qt_x11Data needs to be consistent across the header and the implementation file, and this includes it's visibility.  This error and patch was found while attempting to get symbol visibility to work with the SunStudio compilers (coming later).
I went with the visibility setting from the implementation file
(qapplication_x11.cpp) because I don't know if this variables is used
outside of libQtGui.so, especially in 3rd-party code.  If it's not
supposed to be used outside of libQtGui.so, then the proper change
would be to remove the Q_GUI_EXPORT from the definition in
qapplication_x11.cpp.

Merge-request: 433
Reviewed-by: Thiago Macieira <thiago.macieira@nokia.com>