Improve DEF file enable/disable mechanism on Symbian
Provide configure flag to enable/disable the use of DEF files on Symbian.
A useful side-effect was that it cleaned up how we control DEF files from
qbase.pri and in WebKit.

-nokia-developer still disables DEF files, as it triggers the autotest
exports, which are not frozen into the DEF files.

Disabling DEF files means that there is no BC with previously released
versions of Qt, so this should only be used for development purposes.

.pro files can specify custom locations for DEF files by setting
defFilePath.

Task-number: QTBUG-6556
Reviewed-by: Jason Barron