QListView::setAlternatingRowColors is not working
QS60Style does not really support alternating line/cell backgrounds.
This is due to that no such thing exists on native side and we are
using a graphical theme image as list/table background.

For style, we'll use QPalette::Light with some transparecy to achieve
this. Light is guaranteed to have good contrast with theme background
as it is derived from QPalette::Button, which is calculated from
button graphics, which *should* be used on top of theme background.

Task-number: QTBUG-5980
Reviewed-by: Alessandro Portale