Add a sanity check to verify that the system() command works.
Pre 2.6.23 Linux kernels have a limit on the environment size that
can be passed to a forked process. We quite easily overstep this
boundary when building big projects on Symbian, and since we depend
on running the system() command, this causes the build to fail. This
commit tests that system() can be successfully run.