Cocoa: Usage of QMacCocoaAutoReleasePool makes CPU peak
The reason was that QMacCocoaAutoReleasePool included a
a call to NSApplicationLoad(). This call should only be
made for carbon based application anyway, so we just ifdef
it out (event how clumsy the placing of the call is). The
CPU problem came because after the call, [NSApp isRunning]
would return true, an as such, confuse the event dispatcher
later on.

Reviewed-by: MortenS