Qt/Cocoa Event Dispatcher Problem in modal dialogs
The problem is that we didn't check if the event dispatcher was
interrupted before starting to wait for more events. This patch will do
the interrupt test after processing modal session events, and just
before starting to wait. This will fix applications that expects e.g an
event loop to exit immidiatly upon a signal from a timer (without the
need for the user to generate e.g. a mouse event to stop the wait).

Task-number: QTBUG-7503
Reviewed-by: cduclos