Drawing fake buttons using QMacStyle+QStyleOptionViewItemV4 lead to crash.
The problem here is the fact that we "trusted" the user when the widget
type was specified. The fix consists in checking the result of the
conversions, if the conversion was successful (i.e. the widget was of
the type specified by the user) then we proceed as usual. If the
conversion was not successful (i.e. wrong widget type) then we ask the
style for a sensible size.
I modified this for QPushButton, QSlider and QToolButton.

Task-number: qtbug-7522
Reviewed-by: jbache