Remove unneeded resources from QtGui
QtGui contains loads of resources. 337,160 bytes just for qstyle.
And since this is nearly exclusively full of PNG files,
it won't really be compressed in the qt.sis package.
And it will eat unneccessary Ram.

This patch removes about 260,000 bytes of resources. We only want
the 32-pixel big icons. No 16- and 129- pixel versions.

Task-number: QTBUG-7277
Reviewed-by: Sami Merila

	modified:   src/gui/kernel/qapplication.cpp
	new file:   src/gui/styles/qstyle_s60.qrc
	modified:   src/gui/styles/styles.pri