Cocoa: qaccessibility test crashes
The reason is that the test first creates a menu bar with an exit item
The exit item gets merged into the appmenu (together with a QAction)
Then we destroy the menu bar, but leave the exit item in the appmenu untouched.
Then we create a new menubar _without_ an exit item
macUpdateMenubar will then, at one point, try to access the old exit item's
QAction, wich is destroyed a long time ago; crash!
This patch will make sure we clear out all actions in the menu bars appmenu
when the menubar gets destroyed.

Reviewed-by: Prasanth