Mac: Calling showFullScreen() then showNormal() on a widget results in top menu hiding.
The problem here was the way we entered Full Screen Mode. We were using
"kUIModeAllSuppressed" which does the following according to the manual:
kUIModeAllSuppressed
All system UI elements (including the menu bar) are hidden. However,
these elements may automatically show themselves in response to mouse
movements or other user activity.
I changed it to "kUIModeAllHidden", which does the following:
All system UI elements (including the menu bar) are hidden.

To prevent a change of behavior I added the following option to the
SetSystemUIMode:

kUIOptionAutoShowMenuBar
This flag specifies that the menu bar automatically shows itself when
the user moves the mouse into the screen area that would ordinarily be
occupied by the menu bar.
Only valid for the presentation mode kUIModeAllHidden.

Task-number: QTBUG-7625
Reviewed-by: Richard Moe Gustavsen