Improved the cldr parser for QLocale.
I believe now we parse the CLDR properly. If a data is not present in a file,
we remove the last part of the filename and try again. And we should do the
same if a file is an alias to some other file. For example az_AZ.xml is an
alias to az_Latn_AZ.xml which itself falls back to az_Latn.xml and then to
az.xml

Reviewed-by: Frans Englich