Fix QDirIterator auto test
The test depended on being run inside the source directory, because of
a dummy file hierarchy present there.

With these changes the needed hierarchy is generated on the fly and
cleaned up at the end of the test.

Also removed garbage at the end of tst_QDirIterator::stopLinkLoop.
Cleanup is done in test destructor.