Shortening and fixing previous fix for QTBUG-6371
Shortening: Use much mor code from CommonStyle
Fixing: Subtract PM_Layout[Left|Right]Margin from desktop width.

This fix is anyways not relly useful. It heals some Ui designs with
sizeconstrains which do not fit into mobile screens. But it will fail
In following situations:

Widget, and a too-long combobox in a horizontal Layout. The combobox
would be cropped to Desktop width but the widget(s) left and/or right
of the Combobox are not considered. So, that the Ui still is too wide.

The correct fix will be for set the size policy of the combobox to
something less greedy.

Task-number: QTBUG-6371
Reviewed-by: Sami Merila

	modified:   gui/styles/qs60style.cpp