Fixed `make test' for xmlpatternsxqts.
As far as I can tell, setting lib.file here is unnecessary, and it
happens to break `make test'.

The reason is that, on the first pass, qmake will generate Makefile.lib;
then, when `make' is run, it will generate Makefile, and any additional
parameters originally passed to qmake will be omitted.