Carbon: crash problem in QWidgetPrivate::hide_sys
The crash occurs because we at one point try to assign a widget that
is marked as 'about to be deleted' to a QPointer, after clearguards
has been called. The correct fix is to avoid such an assignment in
the first place.

Task-number: QTBUG-4227
Reviewed-by: Prasanth
(cherry picked from commit f842ec12706b70f94ab5f634dc2aa025ba2cf3f2)