Fixed reading draft data from CLDR.
When draft resolution is specified we should accept it and all values "above"
it. For example if we are searching for a group separator with a "contributed"
resolution, we should also accept data that is "approved".

In theory we should only accept data with "contributed" or "approved"
resolutions and should not trust other resolutions. However in order not to
introduce data regressions to QLocale we will only handle drafts for new data.

Reviewed-by: Frans Englich