Don't crash when running Qt on KDE with Oxygen style.
It crashed on application exit because of the order of Q_GLOBAL_STATICs
are not defined if the global statics are in different modules. Thus,
the globalRedirectionAtomic was destroyed first, then the
K_GLOBAL_STATIC(OxygenHelper, ...) got destroyed. Since the
OxygeneHelper contained a QCache of pixmaps it would eventually call
~QPixmap() and reference the destroyed globalRedirectionAtomic....

Task-number: QTBUG-7734
Reviewed-by: gunnar