Fixed text rendering via scaled QPainter on rasterPaintEngine / Symbian
Text rendering in Qt for Symbian looks very bad on a scaled QPainter.

Instead of simply using a Symbian font with the right size it followed
the code path that is usually used for "tricky transformation":
Use the glyph outlines and rasterize them in Qt as QPainterPath.
That's slow and especially ugly on Symbian since we do not (yet)
get the glyph outlines from Symbian.

This patch adds QFontEngineS60::setFontScale() so that QFontEngineS60
has one additional native font handle with the scaled size. The raster
paint engine will call that function in case that the painter is scaled.
Since we already have QFontEngineS60 specific code in the raster paint
engine, this patch did not have to add #ifdefs.

http://bugreports.qt.nokia.com/browse/QTBUG-7614

Task-number: QTBUG-7614
Reviewed-by: Jason Barron

	modified:   gui/painting/qpaintengine_raster.cpp
	modified:   gui/text/qfontengine_s60.cpp
	modified:   gui/text/qfontengine_s60_p.h
(cherry picked from commit 5a64033f89e07505bb77d3c172597b8df3cbd762)