Implemented pkg generator for the symbian/linux-armcc mkspec.
It reuses almost everything from the original pkg generator, but the
code had to be restructured to accomodate for the new makefile
generator. In addition a few new codepaths and profile changes had to
be introduced to cope with the fact that this mkspec puts files in
different places than Symbian does (inside $QTDIR/lib instead of
$EPOCROOT/...).

This enables the pkg generator to generate correct pkg files for Qt
itself, but applications will not work yet because they also require
generated rss files.