QSortFilterProxyModel: Fix dynamic sorting when severals rows are added.
We need to update the sorting column in clear_mapping.
But the problem is that updating the source column create a mapping
(because it uses mapToSource). So we need to clear that wrong mapping again

In _q_sourceLayoutChanged there was an uneeded special case.
The case where the source_sort_column might change was never triggered
because of that. So now that the special case has been removed, we need
to do as in clear_mapping.

The test cover the change in clear_mapping and in _q_sourceLayoutChanged

The problem was shown in the Kopete: https://bugs.kde.org/show_bug.cgi?id=199850

Reviewed-by: Thierry
Task-number: QTBUG-7537
(cherry picked from commit 7b96944a02991e404cdab61e24182b90f5849120)