Qt applications not responding to WM_GETHOTKEY on windows.
On windows when an application has been started through a shortcut link,
(either by double-clicking or by pressing the shortcut key assigned in
the link) the STARTUPINFO contains the hotkey associated to the
application. This should be set on the toplevel using WM_SETHOTKEY, so
that it behaves like native applications.

Task-number: QUTBUG-3426
Reviewed-by: Jens Bache-Wiig
Reviewed-by: bnilsen