Fix crash when multiple screens are reported from HAL.
The support for multiple screens is not implemented yet in Qt on the
Symbian platform, so there is really no need to query the HAL for the
value since the additional screens can't be used yet anyway. The crash
here occured when the HAL returned more than 1 screen, but the arrays
were resized to contain one element. The loop in resizeEvent() was
iterating past the end of the array because it thought there were more
screens than the arrays did.

Reviewed-by: axis