Crash when closing any top-level widget on Symbian.
The window surface must be deleted while the window is valid because on
some graphics systems a notifcation is sent to the window when the
surface is released.

This fix is also an optmization as we no longer process any backing
store requests while deleting the top-level. Previously it would
handle requests from the window itself and all its children.

Task-number: QT-2513
Reviewed-by: jbarron