Cocoa: Tool Windows steal focus from the main window
The example that shows this off is tools/inputpanel. This example is
strangely written, and calls activateWindow on the main window in an
activate callback to another window. So a bit recursion will happend,
and this caused troubble for the cocoa port. In addition, in Qt several
windows can be active at the same time, expecially tool windows. This
means that we need to be careful when returning early from the
becomeKeyWindow methods for NSWindow. This patch fixes up on this stuff.

Task-number: QTBUG-6002
Reviewed-by: msorvig