Flush the WSERV command buffer after deleting a surface.
For graphics systems that use EGL surfaces in the backing store
destroying the surface does not guarantee that the memory is
immediately freed because this command does not cause a flush. This
implies that a manual flush is instead needed. We do this in 2 places;
the first is when the surface is destroyed due to a visibility changed.
The second case is just after the window has been destroyed. At this
point the backing store has already been deleted so the deletion of
both the surface and window can happen atomically in WSERV.

Task-number: QT-2506
Reviewed-by: Iain
(cherry picked from commit 91e236022acd79dfbc4aef9e30edb4d1aeb2685c)