Initial support for S60 softkey icons.
The logic for resolving the icon size in landscape mode for S60 5.0
(5800XM) is somewhat fragile, but better way is not yet know.

Also the 50% transparent mask what CBA implementation tries to create
for pressed down CB Abuttons fails for some reason. When button is
pressed down there are drawing artifacts in softkey images.

These issues will be tried to resolve with later commits

Task-number: QTBUG-7314

Review-By: Sami Merila