Fix QtScript debugger hang/crash issue with updating locals model
Since the model is updated lazily, it's possible that a new model
has been set on the view (and the old model deleted) before the old
model has finished refreshing. This was discovered while holding
down the F11 key (Step Into) when debugging plasma.js in the
script/context2d example, which causes many events (and hence
asynchronous model updates) to trigger rapidly.