Fixed casual crash in initializeDb (Symbian)
dynamic_cast should be used carefully with rvct before version 4.

We found that out after the a dynamic (down)-cast from CFont* to
CFbsFont* in QtGui suddenly failed.

We test for (TypeUid() == KCFbsFontUid), before doing the
cast. So, a static_cast is safe in this case. Also the other
two dynamic_casts in qfontdatabase_s60.cpp were changed to
static_cast.

http://bugreports.qt.nokia.com/browse/QTBUG-7963

Task-number: QTBUG-7963
Reviewed-by: lars

	modified:   src/gui/text/qfontdatabase_s60.cpp