Added native key support to QSharedMemory API.
Methods setNativeKey() and nativeKey() were added to
QSharedMemory API. Shared memory's native key is returned
by nativeKey() and it is set with either setKey() or
setNativeKey(). setKey() leads to a native key that is
platform independent while setNativeKey() directly sets
the native key without any mangling.

When using setNativeKey(), key() returns a null string
and shared memory's system semaphore is not set. This
means that is up to the user to define a such protection
mechanism (i.e. lock() can't be used on native keys).

QSharedMemory tests were updated.

Merge-request: 1497
Reviewed-by: Benjamin Poulain <benjamin.poulain@nokia.com>
Reviewed-by: Andreas Aardal Hanssen <andreas.aardal.hanssen@nokia.com>