Fixed warnings and crash when painting graphics effects outside scene.
The problem was that when an item had a QGraphicsEffect on itself and
that item was outside of the sceneRect, the cached pixmap of the item
was just an empty pixmap (when using DeviceCoordinates mode). Therefore
the effect filter was trying to paint into an unformatted and empty
image. Now, paint() for all pixmap filters just return immediatly when
the pixmap is empty.

Task-number: QTBUG-5358
Reviewed-by: sroedal